Create RecursivelyRemoveNodeInLinkedlist.java
Problem: Given internal access to a Singly-Linked List with Comparable data, where all of the data is guaranteed to be in sorted order, remove all duplicates from the list so that there is at most one of each data. This task should be done in  time and  space (there is space overhead for using recursion in the first place, but by  space here, we are referring to external memory such as creating new nodes, data, etc.).